feat: refactor route
